requirements.txt
drf_yasg==1.20.0
Django==2.2.10
djangorestframework==3.12.4
